Implement chain and more run-options
Options need to be handled in an ordered way. It would be too much work
to force only one parameter at a time, such as

`{ withError: string; withErrors?: undefined } | { withError?: undefined; withErrors: string[] }`

as that would be an explosion of combinations. Therefore we just handle
the combinations that make sense, allowing you to specify both one and
many errors, and all will be present in the final error message. Of
course, this requires defining some arbitrary semantics, such as
`withDefault` overriding any error - which could just as well be the
other way around.