Implement choice and fix iterator copy bug
Choice is implemented somewhat correctly here, but the edgecase of no
parsers needs a better message; maybe no message? This isn't a
programming mistake as I first assumed; it's the end of the recursion,
so it is always reached.

Iterator needs to be copied at current index (number of iterated elemnts
so far), or else obviously, the new iterator starts over, reading from
the beginning, which is not what we want.