Implement repeat and many
This uncovered lots of bugs. `choice` wasn't advancing the state of the
iterator properly because of overly defensive copying. Also, we were
moving the state too much by not copying in the right places. It went
wrong in both directions!

Next I will try factoring it to a bindFailure kind of thing, to at
least contain the errors in one place.