Refactor `bindError` to be less error prone
Also implement `option`. It's kind of interesting how `option` returns a
list with one element, or an empty list, depending on whether it
succeeded. The nice thing with this is that you can assign a default
value when destructuring, kind of like

```
const [result = 'default'] = p.option(parser);
```

As long as you're not fooled to thinking it can return more than one
element.

The `bindError` function takes the list of error messages produced so far,
and the current state of the stream (except not moved, since it has
failed), and can do whatever it wants with it - be it parse stuff
successfully or throw, itself.